Move spider settings tests
